feat: 연관검색어 처리 훅 shared 파일로 분리
