+++ apps/client/src/pages/search/hooks/use-search-data.ts
@@ -1,4 +1,4 @@
-import { useQueries, useQuery, useSuspenseQuery } from '@tanstack/react-query';
+import { useQuery, useSuspenseQuery } from '@tanstack/react-query';
 
 import { SEARCH_ARTIST_QUERY_OPTION } from '@shared/apis/search/search-queries';
 import { IntendedPerformanceRequest } from '@shared/types/search-reponse';
@@ -16,25 +16,6 @@ export const useSearchArtist = ({ keyword, enabled }: KeywordProps) => {
   });
 
   return { data, isLoading };
-};
-
-export const useRelatedSearch = ({ keyword, enabled }: KeywordProps) => {
-  return useQueries({
-    queries: [
-      SEARCH_ARTIST_QUERY_OPTION.SEARCH_RELATED_KEYWORD(keyword, enabled),
-      SEARCH_PERFORMANCE_QUERY_OPTION.SEARCH_RELATED_PERFORMANCES(
-        keyword,
-        enabled,
-      ),
-    ],
-    combine: (results) => ({
-      data: {
-        relatedArtists: results[0].data,
-        relatedPerformances: results[1].data,
-      },
-      isLoading: results.some((r) => r.isLoading),
-    }),
-  });
 };
 
 export const usePerformanceTypeAnalysis = ({
